Test: 리뷰 상세 API 및 Selector 테스트 코드 수정 #37
- Review Selector 오타 수정으로 테스트 코드 동일 수정
- Review 상세 API response값 수정으로 인한 테스트 코드 수정
Related to #37
